core-node/quartapp/requirements.txt
quart 
pandas 
numpy 
matplotlib 
seaborn 
scikit-learn 
ipykernel 
asyncio-redis 
redis 
Werkzeug 
nest_asyncio 
Flask-Cors 
asyncio 
hypercorn 
pytest-asyncio 
aiohttp
argon2-cffi
Flask-Security-Too
mongoengine